WIP: most tests run, but need to fix schema for darks
