+++ xpdan/pipelines.py
@@ -102,3 +102,34 @@ def integration_pipeline(img_stream,
                        mask_stream=mask_stream,
                        **integration_kwargs)
     yield from iq_stream
+
+
+def db_integrate(img_hdr,
+                 glbl=an_glbl,
+                 **kwargs
+                 ):
+    exp_db = glbl.exp_db
+    an_db = glbl.an_db
+    img_stream = exp_db.restream(img_hdr, fill=True)
+
+    # Replace 'hdr's with 'stream's in kwargs
+    for key, db in [('dark_hdr', exp_db),
+                    ('detector_calibration_hdr', an_db),
+                    ('mask_hdr', an_db),]:
+        new_key = key.replace('hdr', 'stream')
+        # If the header is None or doesn't exist the stream is None
+        if kwargs.get(key, None) is None:
+            kwargs[new_key] = None
+        # Otherwise restream the header
+        else:
+            kwargs[new_key] = db.restream(key, fill=True)
+            kwargs.pop(key)
+
+    kwargs.update(glbl=glbl)
+    p = integration_pipeline(img_stream, **kwargs)
+    for n, d in p:
+        if n == 'start':
+            rv = d['uid']
+        pass
+
+    return rv
